Refactor interpreter to use arbitrary precision arithmetic (BigInteger/BigDecimal)
- Split `Num.java` into `NumOps.java` (arithmetic strategies) and `NumUtils.java` (shared utilities) to improve modularity and maintainability.
- Updated `Num.ops(Object)` dispatch logic to return `BIGINT_OPS` for all integer types and `BIGDECIMAL_OPS` for all floating-point types. This forces the interpreter to perform calculations using `BigInteger` and `BigDecimal`, preventing unboxing to primitive types and ensuring arbitrary precision.
- Patched `NumUtils.bitOpsCast` to handle `BigInteger` inputs by converting them to `BigInteger` (upscaling primitives), preventing runtime crashes in bitwise operations and providing arbitrary precision bitwise logic.
- Updated `Num.java` bitwise methods (`and`, `or`, `xor`, etc.) to use `BigInteger` arithmetic and return `BigInteger`.
- Updated `Builtin.java` bitwise wrappers to return `Number` instead of `Long`.
- Updated unit tests (`ControlTest`, `EvalTest`, `RTTest`) to assert `BigInteger` results instead of `Long`, reflecting the new semantic behavior.
- Verified changes with `mvn clean compile` and comprehensive test runs (`NumTest`, `hara.lang.data.*Test`).

+++ src/main/java/hara/lang/base/NumUtils.java
@@ -128,13 +128,13 @@ public class NumUtils {
         return -x;
     }
 
-    public static long bitOpsCast(Object x) {
+    public static BigInteger bitOpsCast(Object x) {
 		Class<? extends Object> xc = x.getClass();
 
 		if (xc == Long.class || xc == Integer.class || xc == Short.class || xc == Byte.class)
-			return (long) ((Number) x).longValue();
+			return BigInteger.valueOf(((Number) x).longValue());
         else if (x instanceof BigInteger)
-            return ((BigInteger) x).longValue();
+            return (BigInteger) x;
 		// no bignums, no decimals
 		throw new IllegalArgumentException("bit operation not supported for: " + xc);
 	}
